diagnose the case where there are too many combinators in a selector
Because the implementation uses integers as bit fields to represent
sets, there can be only 63 (or 31 if your perl uses 32-bit integers)
child/sibling combinators (i.e. >, ~, +) in a row (i.e. without an
intervening descentant combinator (space)).

I don't expect this limit to be much of a problem in practice, so just
throw an error for now.